move AWS_BACKBONE_HOSTNAME to AwsConfiguration, make public (#5977)
* move AWS_BACKBONE_HOSTNAME to AwsConfiguration, make public
