+++ projects/batfish/src/test/java/org/batfish/representation/aws/AwsConfigurationNatGatewayTest.java
@@ -1,7 +1,7 @@
 package org.batfish.representation.aws;
 
 import static org.batfish.common.util.isp.IspModelingUtils.INTERNET_HOST_NAME;
-import static org.batfish.representation.aws.AwsConfigurationTestUtils.AWS_BACKBONE_HOSTNAME;
+import static org.batfish.representation.aws.AwsConfiguration.AWS_BACKBONE_HOSTNAME;
 import static org.batfish.representation.aws.AwsConfigurationTestUtils.getTcpFlow;
 import static org.batfish.representation.aws.AwsConfigurationTestUtils.testSetup;
 import static org.batfish.representation.aws.AwsConfigurationTestUtils.testTrace;
