+++ projects/batfish/src/test/java/org/batfish/representation/aws/AwsConfigurationVpcEndpointGatewayTest.java
@@ -1,7 +1,7 @@
 package org.batfish.representation.aws;
 
+import static org.batfish.representation.aws.AwsConfiguration.AWS_BACKBONE_HOSTNAME;
 import static org.batfish.representation.aws.AwsConfiguration.AWS_SERVICES_GATEWAY_NODE_NAME;
-import static org.batfish.representation.aws.AwsConfigurationTestUtils.AWS_BACKBONE_HOSTNAME;
 import static org.batfish.representation.aws.AwsConfigurationTestUtils.getTcpFlow;
 import static org.batfish.representation.aws.AwsConfigurationTestUtils.testBidirectionalTrace;
 import static org.batfish.representation.aws.AwsConfigurationTestUtils.testSetup;
